pyproject.toml
[project]
name = "walkie-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "deepagents>=0.3.9",
    "elevenlabs>=2.33.1",
    "faster-whisper>=1.0.0",
    "google-cloud-speech>=2.36.0",
    "langchain-openai>=1.1.7",
    "pyaudio>=0.2.14",
    "pydub>=0.25.1",
    "python-dotenv>=1.2.1",
    "scipy>=1.17.0",
    "silero-vad>=6.2.0",
    "sounddevice>=0.5.5",
    "walkie-sdk",
]

[tool.uv.sources]
walkie-sdk = { git = "https://github.com/EIC-Robocup-2026/walkie-sdk.git" }
